feat(psychology): 주식 자동 매매 페이지 상단 서브탭에 심리 분석 추가
기존에는 좌측 사이드바(InvestmentLayout)에서만 노출되던 심리 분석을
대시보드의 주식 자동 매매 화면(InvestmentTab) 상단 가로 서브탭에도 추가.
자동매매 활성화된 사용자가 같은 화면 내에서 자유롭게 사용 가능.

- src/components/Investment/Psychology/PsychologyContent.tsx (신규):
  PsychologyPage 인라인 컴포넌트를 별도 파일로 분리 (InvestmentTab/PsychologyPage 양쪽 재사용)
- src/components/Investment/InvestmentTab.tsx:
  SUB_TABS에 'psychology' 항목 추가 (자동매매 ↔ 포트폴리오 사이), Users 아이콘 사용
- src/app/investment/psychology/page.tsx:
  분리된 PsychologyContent를 wrap만 (기존 라우트 유지)
